chore(deps): updating dependencies
Updating dependencies? What a simple change you say....but no, this is
android where updating dependencies means also updating gradle, and
android, and kotlin and then prior building code suddenly breaks...
Where's that chicken for the sacrifice...the gods are angry and a
sacrificial farm animal is the only way to get your app to build
again...or stack overflow...you're safe for now mr. chicken